Added TimeField input
Before the commit, there were two types of date/time fields: DateTimeField and DateField.
Using model.TimeField would result in a regular text field which would be easily broken by locale (e.g: 17:00 would be converted to 5 p.m., which is not a correct TimeField input)
So I basically made a copy of DateInput and modified it. Current time format is HH:MM (12:23) because seconds are something that we don't really enter manually.